+++ CHANGELOG.md
@@ -9,8 +9,12 @@ Release Notes
 
     ```diff
      final class DatabasePool {
-    -    init(path: String, configuration: Configuration = default) throws
-    +    init(path: String, configuration: Configuration = default, maximumReaderCount: Int = default) throws
+    -    init(path: String, configuration: Configuration = default, maximumReaderCount: Int = default) throws
+    +    init(path: String, configuration: Configuration = default) throws
+     }
+     struct Configuration {
+    +    var maximumReaderCount: Int = default
+     }
     ```